feat: Implement road connection bonus for trade route shares and remove TradeShipExecution class
